faet: Исправлена логика добавления и удаления транзакции в офлайн режиме
