+++ Expense Tracker/Services/Categories/CategoriesStorage/CategoriesStorageProtocol.swift
